requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
transformers==4.42.4
torch==2.3.1
Pillow==10.4.0
requests==2.32.3
pydantic==2.8.2
